chore: add neutral workflow progress builder
