+++ ui/application/__init__.py
@@ -113,6 +113,7 @@ from .local_first_progress_facts import (
     current_local_first_last_sync_date,
     runtime_state_with_local_first_output_path,
 )
+from .workflow_progress import build_workflow_progress_from_facts
 from .workflow_progress_facts import (
     WorkflowProgressFacts,
     build_workflow_progress_facts_from_runtime_state,
@@ -270,6 +271,7 @@ __all__ = [
     "current_local_first_last_sync_date",
     "build_wizard_filter_description",
     "build_workflow_progress_facts_from_runtime_state",
+    "build_workflow_progress_from_facts",
     "build_wizard_progress_facts_from_runtime_state",
     "build_wizard_progress_from_facts",
     "build_wizard_progress_from_facts_and_settings",
